services/backend/requirements.txt
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
asyncpg==0.29.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
DateTime==5.4
distro==1.9.0
fastapi==0.110.0
feedparser==6.0.11
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
Mako==1.3.2
MarkupSafe==2.1.5
miniflux==1.0.1
numpy==1.26.4
openai==1.13.3
pandas==2.2.1
passlib==1.7.4
pgvector==0.2.5
pinecone-client==3.1.0
pycparser==2.21
pydantic==2.6.3
pydantic-settings==2.2.1
pydantic_core==2.16.3
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
regex==2023.12.25
requests==2.31.0
setuptools==69.1.1
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.28
SQLAlchemy-Utils==0.41.1
starlette==0.36.3
tiktoken==0.6.0
tqdm==4.66.2
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.27.1
zope.interface==6.2
validators==0.22.0